+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Bug Fixes:
     * runtime: INPUT statement fixed
     * build system: makefile portability enhancements
     * compiler: fix BYREF recursive argument passing
+    * compiler: fix BYREF argument assignment
     * compiler: fix IF/ELSE namespaces
 
 ## 0.7.3